convert ot intgere

+++ db/base.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "users" (
-  "id" bigserial NOT NULL,
+  "id" serial NOT NULL,
   "username" character varying(255) NOT NULL,
   "password" text NOT NULL,
   "salt" text NOT NULL,
@@ -21,7 +21,7 @@ END ;$$
 CREATE TRIGGER update_user_timestamp BEFORE UPDATE ON users FOR EACH ROW EXECUTE PROCEDURE update_timestamp();
 
 CREATE TABLE "temp_codes" (
-  "user_id" bigint NOT NULL REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE,
+  "user_id" integer NOT NULL REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE,
   "hash" character varying(255) NOT NULL,
   "type" character varying(255) NOT NULL,
   "created_at" timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -29,10 +29,10 @@ CREATE TABLE "temp_codes" (
 );
 
 CREATE TABLE "rooms" (
-  "id" bigserial NOT NULL,
+  "id" serial NOT NULL,
   "name" character varying(20) NOT NULL UNIQUE,
-  "player1_id" bigint NOT NULL REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE UNIQUE,
-  "player2_id" bigint REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE UNIQUE,
+  "player1_id" integer NOT NULL REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE UNIQUE,
+  "player2_id" integer REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE UNIQUE,
   "created_at" timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   "updated_at" timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (id)
@@ -45,7 +45,7 @@ GRANT ALL ON rooms TO velioo;
 GRANT ALL ON users_id_seq TO velioo;
 
 CREATE TABLE "gameplay_statuses" (
-  "id" bigint NOT NULL,
+  "id" integer NOT NULL,
   "name" TEXT NOT NULL,
   PRIMARY KEY (id)
 );
@@ -53,7 +53,7 @@ CREATE TABLE "gameplay_statuses" (
 INSERT INTO gameplay_statuses (id, name) VALUES (1, 'In progress'), (2, 'Finished');
 
 CREATE TABLE "boards" (
-  "id" bigserial NOT NULL,
+  "id" serial NOT NULL,
   "board_matrix_json" text NOT NULL,
   "board_data_json" text NOT NULL,
   "created_at" timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -69,15 +69,15 @@ GRANT ALL ON boards TO velioo;
 GRANT ALL ON boards_id_seq TO velioo;
 
 CREATE TABLE "games" (
-  "id" bigserial NOT NULL,
-  "room_id" bigint UNIQUE,
-  "player1_id" bigint NOT NULL REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE,
-  "player2_id" bigint NOT NULL REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE,
+  "id" serial NOT NULL,
+  "room_id" integer UNIQUE,
+  "player1_id" integer NOT NULL REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE,
+  "player2_id" integer NOT NULL REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE,
   "data_json" text NOT NULL,
   "room_data_json" text NOT NULL,
-  "status_id" bigint NOT NULL REFERENCES gameplay_statuses(id) ON UPDATE CASCADE,
-  "winning_player_id" bigint REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE,
-  "board_id" bigint NOT NULL REFERENCES boards(id) ON UPDATE CASCADE,
+  "status_id" integer NOT NULL REFERENCES gameplay_statuses(id) ON UPDATE CASCADE,
+  "winning_player_id" integer REFERENCES users(id) ON DELETE CASCADE ON UPDATE CASCADE,
+  "board_id" integer NOT NULL REFERENCES boards(id) ON UPDATE CASCADE,
   "created_at" timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   "updated_at" timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   "finished_at" timestamp,
